requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.7
django-autoslug==1.9.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.15.0
idna==3.3
kombu==5.2.4
packaging==21.3
Pillow==8.3.2
prompt-toolkit==3.0.30
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
